add unittests
